Add an integration test suite
This changes adds the first set of integration tests for the NFM agent.  A
Dockerfile is introduced to define an integ test container image.  Within that
container all scripts are invoked that match the path
`test-data/integration-test*`.  The first of those tests included in this
change performs a few HTTP requests between client and server within the local
host, and validates the NFM agent reports the appropriate stats.